fetch/stale-while-revalidate/stale-css.tentative.html

<!DOCTYPE html>
<!---
Tentative test against:
https://github.com/whatwg/fetch/pull/853
-->
<meta charset="utf-8">
<title>Tests Stale While Revalidate works for css</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>

async_test(t => {
  window.onload = t.step_func(() => {
    step_timeout(() => {
      assert_equals(window.getComputedStyle(document.body).getPropertyValue('background-color'), "rgb(0, 128, 0)");
      var link2 = document.createElement("link");
      link2.onload = t.step_func_done(() => {
        assert_equals(window.getComputedStyle(document.body).getPropertyValue('background-color'), "rgb(0, 128, 0)");
      });
      link2.rel = "stylesheet";
      link2.type = "text/css";
      link2.href = "stale-css.py";
      document.body.appendChild(link2);
    }, 0);
  });
}, 'Cache returns stale resource');

var link = document.createElement("link");
link.rel = "stylesheet";
link.type = "text/css";
link.href = "stale-css.py";
document.body.appendChild(link);
</script>
</body>
